# specs/platform/api.yaml
...
openapi: 3.1.0
info:
  title: Wizard platform-api
  version: 1.0.0
  description: 'Нормативный контракт apps/platform-api (backlog M0-15, M1-02..M1-04, M1-07, M2-07..M2-09). Каждая операция помечена x-milestone; контрактный тест M0-15 покрывает все операции с x-milestone ≤ текущей вехи. Правила: все тела — JSON (кроме SSE и /files); все пользовательские тексты в ответах — на русском; идентификаторы — uuid; время — RFC 3339 UTC; кредиты в API — число с точностью 0.001 (= милликредит, billing.yaml#credit.unit), округление до 0.1 — только в UI. Тексты, пришедшие от LLM (сообщения агентов, карточка, message_ru гейтов), клиент MUST показывать только как текст или санированный markdown без raw HTML (L3-17). Обращение к объекту чужой организации (system, run, revision, file, events, export) → 404 NOT_FOUND, не 403 (контрактный IDOR-тест M0-15/M1-02). Мутирующие POST принимают заголовок Idempotency-Key (≤ 64 символа): повтор с тем же ключом и тем же телом в течение 24 ч MUST вернуть исходный ответ; с другим телом — 422 IDEMPOTENCY_MISMATCH. Авторизация по вехам — x-auth. Права ролей организации — x-roles (owner ⊃ editor ⊃ viewer, product.yaml#decisions.D11_team).

    '
  x-auth:
...
    M2: '+ staff: users.is_staff=true, MFA (TOTP) обязательна: без users.mfa_enrolled_at или без TOTP-подтверждения за последние 12 ч /admin/* → 403 MFA_REQUIRED. В бете staff — учётная запись основателя (F7); все действия /admin/* пишутся в staff_audit_log.'
servers: [{url: 'http://localhost:4000/api/v1', description: M0–M1 локально}]
security: [{devUser: []}, {session: []}]
paths:
...
  /systems/{id}/publish:
    parameters: [{$ref: '#/components/parameters/systemId'}]
    post:
      operationId: publish
      x-milestone: M1
      x-roles: [owner]
      summary: 'Публикация ревизии в prod (workflows.yaml#workflows.publish). M2+ требует карту РФ и прохождение G2. M2-09: при WIZARD_FOUNDER_REVIEW=on и orgs.require_founder_review первая публикация системы и ревизия с новыми полями ПДн ждут одобрения основателя (прогон → GATES_FAILED с текстом abuse.yaml#messages_ru.review, затем 403 FOUNDER_REVIEW_PENDING до approve)'
      parameters: [{$ref: '#/components/parameters/idempotencyKey'}]
      requestBody: {required: true, content: {application/json: {schema: {type: object, required: [revision], additionalProperties: false, properties: {revision: {type: integer}, confirmDiff: {type: boolean, const: true, description: Пользователь видел diff}}}}}}
      responses:
        '202': {description: Прогон publish поставлен в очередь, content: {application/json: {schema: {type: object, properties: {run: {$ref: '#/components/schemas/Run'}}}}}}
        '402': {description: 'PLAN_LIMIT (лимит prod-систем тарифа, billing.yaml#plans.enforcement)', content: {application/json: {schema: {$ref: '#/components/schemas/Error'}}}}
        '403': {$ref: '#/components/responses/Error'}
        '409': {$ref: '#/components/responses/Error'}
        '422': {$ref: '#/components/responses/Error'}
...
  /orgs:
    get: {operationId: listOrgs, x-milestone: M1, responses: {'200': {description: Организации пользователя, content: {application/json: {schema: {type: object, properties: {items: {type: array, items: {$ref: '#/components/schemas/Org'}}}}}}}}}
    post:
      operationId: createOrg
      x-milestone: M1
      summary: 'Создатель становится owner; стартовый грант Free (billing.yaml#plans.free.grants). M2-09: при WIZARD_REGISTRATION=invite новая org владельца пилотной org — plan pilot без кредитов и с require_founder_review=true; без своей пилотной org — 403 FORBIDDEN «На пилоте новые организации создаёт команда Wizard…»'
      requestBody: {required: true, content: {application/json: {schema: {type: object, required: [name], properties: {name: {type: string, minLength: 1, maxLength: 120}}}}}}
      responses: {'201': {description: Создана, content: {application/json: {schema: {$ref: '#/components/schemas/Org'}}}}}
  /orgs/{orgId}:
    parameters: [{$ref: '#/components/parameters/orgId'}]
    get: {operationId: getOrg, x-milestone: M1, x-roles: [viewer], responses: {'200': {description: Организация, content: {application/json: {schema: {$ref: '#/components/schemas/Org'}}}}}}
    patch:
      operationId: updateOrg
      x-milestone: M1
      x-roles: [owner]
      requestBody: {required: true, content: {application/json: {schema: {type: object, properties: {name: {type: string, maxLength: 120}}}}}}
      responses: {'200': {description: Обновлена, content: {application/json: {schema: {$ref: '#/components/schemas/Org'}}}}}
  /orgs/{orgId}/members:
    parameters: [{$ref: '#/components/parameters/orgId'}]
    get: {operationId: listMembers, x-milestone: M1, x-roles: [viewer], responses: {'200': {description: Участники, content: {application/json: {schema: {type: object, properties: {items: {type: array, items: {$ref: '#/components/schemas/Member'}}}}}}}}}
  /orgs/{orgId}/members/{userId}:
    parameters: [{$ref: '#/components/parameters/orgId'}, {name: userId, in: path, required: true, schema: {type: string, format: uuid}}]
    patch:
      operationId: updateMemberRole
      x-milestone: M1
      x-roles: [owner]
      summary: Нельзя оставить организацию без owner → 422 LAST_OWNER
      requestBody: {required: true, content: {application/json: {schema: {type: object, required: [role], properties: {role: {$ref: '#/components/schemas/OrgRole'}}}}}}
      responses: {'200': {description: Обновлено, content: {application/json: {schema: {$ref: '#/components/schemas/Member'}}}}, '422': {$ref: '#/components/responses/Error'}}
    delete: {operationId: removeMember, x-milestone: M1, x-roles: [owner], responses: {'204': {description: Удалён}, '422': {$ref: '#/components/responses/Error'}}}
  /orgs/{orgId}/invites:
    parameters: [{$ref: '#/components/parameters/orgId'}]
    get: {operationId: listInvites, x-milestone: M1, x-roles: [owner], responses: {'200': {description: Активные приглашения, content: {application/json: {schema: {type: object, properties: {items: {type: array, items: {$ref: '#/components/schemas/Invite'}}}}}}}}}
    post:
      operationId: createInvite
      x-milestone: M1
      x-roles: [owner]
      summary: Письмо со ссылкой; TTL 7 дней; лимит участников тарифа → 402 PLAN_LIMIT
      requestBody: {required: true, content: {application/json: {schema: {type: object, required: [email, role], properties: {email: {type: string, format: email}, role: {$ref: '#/components/schemas/OrgRole'}}}}}}
      responses: {'201': {description: Создано, content: {application/json: {schema: {$ref: '#/components/schemas/Invite'}}}}, '402': {$ref: '#/components/responses/Error'}}
  /orgs/{orgId}/invites/{inviteId}: {parameters: [{$ref: '#/components/parameters/orgId'}, {name: inviteId, in: path, required: true, schema: {type: string, format: uuid}}], delete: {operationId: revokeInvite, x-milestone: M1, x-roles: [owner], responses: {'204': {description: Отозвано}}}}
  /invites/{token}/accept:
    parameters: [{name: token, in: path, required: true, schema: {type: string, minLength: 32}}]
    post: {operationId: acceptInvite, x-milestone: M1, summary: email сессии MUST совпадать с email приглашения, responses: {'200': {description: Членство создано, content: {application/json: {schema: {$ref: '#/components/schemas/Member'}}}}, '410': {$ref: '#/components/responses/Error'}}}
  /orgs/{orgId}/settings:
    parameters: [{$ref: '#/components/parameters/orgId'}]
    get: {operationId: getOrgSettings, x-milestone: M1, x-roles: [viewer], responses: {'200': {description: Настройки, content: {application/json: {schema: {$ref: '#/components/schemas/OrgSettings'}}}}}}
    patch:
      operationId: updateOrgSettings
      x-milestone: M1
      x-roles: [owner]
      summary: ruOnly=true — все LLM-вызовы организации только T0 (D2); действует с первого следующего вызова
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/OrgSettings'}}}}
      responses: {'200': {description: Обновлено, content: {application/json: {schema: {$ref: '#/components/schemas/OrgSettings'}}}}}
...
components:
  securitySchemes: {devUser: {type: apiKey, in: header, name: X-Wizard-Dev-User, description: M0; только WIZARD_AUTH_MODE=dev}, session: {type: apiKey, in: cookie, name: __Host-wizard_session, description: 'M1+ (локально по http — wizard_session); + заголовок X-Wizard-CSRF для мутаций'}}
  parameters:
    systemId: {name: id, in: path, required: true, schema: {type: string, format: uuid}}
...
    orgId: {name: orgId, in: path, required: true, schema: {type: string, format: uuid}}
...
    idempotencyKey: {name: Idempotency-Key, in: header, schema: {type: string, maxLength: 64}}
  responses: {Error: {description: Ошибка, content: {application/json: {schema: {$ref: '#/components/schemas/Error'}}}}}
  schemas:
    Error:
      type: object
      required: [code, message_ru]
      additionalProperties: false
      properties:
        code:
          type: string
          description: 'M2-15 (D24_pilot_free): REGISTRATION_INVITE_ONLY — verifyOtp нового email при WIZARD_REGISTRATION=invite без приглашения основателя (CLI pilot invite) и без активного приглашения в org: «Регистрация в Wizard пока только по приглашению…»; PAYMENTS_DISABLED — startCardBinding, changeSubscription, cancelSubscription, createTopup при WIZARD_PAYMENTS=off: «Оплата на пилоте отключена — кредиты начисляет команда Wizard»; LLM_BUDGET_EXHAUSTED — новые interview_turn, build (create/change/fix) и import_table, когда Σ billable cost_rub live-вызовов llm_calls за календарный месяц (Europe/Moscow) ≥ WIZARD_LLM_MONTHLY_CAP_RUB (eval считается отдельно, eval.yaml#live_cadence.budget): «Месячный лимит платформы на работу моделей исчерпан…»'
          x-http-status:
            VALIDATION_FAILED: 400
            UNAUTHORIZED: 401
            OTP_INVALID: 401
            INSUFFICIENT_CREDITS: 402
            PLAN_LIMIT: 402
            FORBIDDEN: 403
            MFA_REQUIRED: 403
            NOT_OWNER: 403
            CARD_BINDING_REQUIRED: 403
            FOUNDER_REVIEW_PENDING: 403
            ORG_SUSPENDED: 403
            REGISTRATION_INVITE_ONLY: 403
            PAYMENTS_DISABLED: 403
            NOT_FOUND: 404
            SYSTEM_LOCKED: 409
            NO_GATE_FAILURE: 409
            RUN_NOT_CANCELLABLE: 409
            RUN_NOT_WAITING_INPUT: 409
            CARD_VERSION_STALE: 409
            NO_CARD: 409
            PREVIEW_NOT_READY: 409
            INVITE_EXPIRED: 410
            EXPORT_LINK_USED: 410
            VERSION_CONFLICT: 412
            PAYLOAD_TOO_LARGE: 413
            UNSUPPORTED_MEDIA_TYPE: 415
            HOST_NOT_ALLOWED: 421
            OPS_INVALID: 422
            IDEMPOTENCY_MISMATCH: 422
            CONSENT_REQUIRED: 422
            LAST_OWNER: 422
            CARD_BINDING_REJECTED: 422
            OPERATOR_NAME_REQUIRED: 422
            OPERATOR_CONTACT_REQUIRED: 422
            OPERATOR_ADDRESS_REQUIRED: 422
            INN_INVALID: 422
            PHONE_LOGIN_PLAN_REQUIRED: 422
            CARD_NOT_RU: 422
            SYSTEM_SUSPENDED: 422
            GATES_FAILED: 422
            DESTRUCTIVE_IN_PROD: 422
            ROLLBACK_TARGET_INVALID: 422
            RATE_LIMITED: 429
            INTERNAL: 500
            LLM_UNAVAILABLE: 503
            LLM_BUDGET_EXHAUSTED: 503
        message_ru: {type: string, description: 'Готовый к показу текст на русском, без стеков и SQL'}
        details: {type: object, description: 'Структурные данные. Для OPS_INVALID — {errors: OpsError[]} (ops.yaml#apply.error_shape; коды — ops.yaml#apply.error_codes, в т. ч. OWNER_ONLY_FIELD); для GATES_FAILED — {reports: GateReport[]}; для INSUFFICIENT_CREDITS — {available, required}; для PLAN_LIMIT — {limit, current, plan}'}
    OrgRole: {enum: [owner, editor, viewer]}
...
    Run:
      type: object
      required: [id, kind, status, createdAt]
      properties:
        id: {type: string, format: uuid}
        systemId: {type: [string, 'null'], format: uuid, description: null только для nightly_eval (db.yaml#runs.system_id)}
        kind: {enum: [interview_turn, build, publish, rollback, import_table, retention, nightly_eval, export]}
        mode: {enum: [create, change, fix, point_edit], description: 'только для build; point_edit — M3'}
        status: {enum: [queued, waiting_lock, running, needs_input, succeeded, failed, cancelled], description: workflows.yaml#run_lifecycle}
        currentStep: {type: [string, 'null']}
        credits: {type: object, properties: {estimate: {type: number}, cap: {type: number}, used: {type: number}}}
        baseRevision: {type: [integer, 'null']}
        resultRevision: {type: [integer, 'null']}
        failure: {type: [object, 'null'], properties: {code: {type: string}, message_ru: {type: string}}}
        lastEventSeq: {type: integer}
        createdAt: {type: string, format: date-time}
        startedAt: {type: [string, 'null'], format: date-time}
        finishedAt: {type: [string, 'null'], format: date-time}
...
    Org: {type: object, properties: {id: {type: string, format: uuid}, name: {type: string}, plan: {enum: [free, pilot, start, business], description: 'pilot — только CLI основателя (billing.yaml#plans.pilot)'}, role: {$ref: '#/components/schemas/OrgRole'}, cardBound: {type: boolean}, paymentsEnabled: {type: boolean, description: 'M2-15: WIZARD_PAYMENTS платформы; false — S-billing скрывает покупку, подписки и привязку карты'}}}
    Member: {type: object, properties: {userId: {type: string}, email: {type: string}, name: {type: [string, 'null']}, role: {$ref: '#/components/schemas/OrgRole'}, joinedAt: {type: string, format: date-time}}}
    Invite: {type: object, properties: {id: {type: string}, email: {type: string}, role: {$ref: '#/components/schemas/OrgRole'}, expiresAt: {type: string, format: date-time}}}
    OrgSettings: {type: object, additionalProperties: false, properties: {ruOnly: {type: boolean, default: false}, buildModelLabel: {type: string, readOnly: true, description: 'Фактическая модель сборки по умолчанию для org с учётом ruOnly, t1Restricted и решения F2 (agents/models.yaml), например «GLM-5.3» или «модели в РФ»; S1/S3/S10 показывают её, а не константу (L4-20)'}, t1Restricted: {type: boolean, readOnly: true, description: 'db.yaml#orgs.t1_restricted (L3-02); ставит платформа/staff'}}}
...